Add Settings modal and All Words option
To spice things up, a new setting has been released which will show all words from the Wordle dictionary instead of just the answers. This increases the dictionary size from 2,300+ to 14,800+ words! Not for the faint of heart. Have fun!

This update also includes dependency upgrades to resolve reported vulnerabilities.